% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.applicationdiscoveryservice_operations.R
\name{associate_configuration_items_to_application}
\alias{associate_configuration_items_to_application}
\title{Associates one or more configuration items with an application}
\usage{
associate_configuration_items_to_application(applicationConfigurationId,
  configurationIds)
}
\arguments{
\item{applicationConfigurationId}{The configuration ID of an application with which items are to be associated.}

\item{configurationIds}{The ID of each configuration item to be associated with an application.}
}
\description{
Associates one or more configuration items with an application.
}
\examples{

}
