% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.applicationdiscoveryservice_operations.R
\name{batch_delete_import_data}
\alias{batch_delete_import_data}
\title{Deletes one or more import tasks, each identified by their import ID}
\usage{
batch_delete_import_data(importTaskIds)
}
\arguments{
\item{importTaskIds}{[required] The IDs for the import tasks that you want to delete.}
}
\description{
Deletes one or more import tasks, each identified by their import ID. Each import task has a number of records that can identify servers or applications.
}
\details{
AWS Application Discovery Service has built-in matching logic that will identify when discovered servers match existing entries that you've previously discovered, the information for the already-existing discovered server is updated. When you delete an import task that contains records that were used to match, the information in those matched records that comes from the deleted records will also be deleted.
}
\section{Accepted Parameters}{
\preformatted{batch_delete_import_data(
  importTaskIds = list(
    "string"
  )
)
}
}

