% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.applicationdiscoveryservice_operations.R
\name{create_application}
\alias{create_application}
\title{Creates an application with the given name and description}
\usage{
create_application(name, description = NULL)
}
\arguments{
\item{name}{Name of the application to be created.}

\item{description}{Description of the application to be created.}
}
\description{
Creates an application with the given name and description.
}
\examples{

}
