% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.applicationdiscoveryservice_operations.R
\name{create_tags}
\alias{create_tags}
\title{Creates one or more tags for configuration items}
\usage{
create_tags(
  configurationIds = list(
    "string"
  ),
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}
\arguments{
\item{configurationIds}{[required] A list of configuration items that you want to tag.}

\item{tags}{[required] Tags that you want to associate with one or more configuration items. Specify the tags that you want to create in a \emph{key}-\emph{value} format. For example:

\code{{"key": "serverType", "value": "webServer"}}}
}
\description{
Creates one or more tags for configuration items. Tags are metadata that help you categorize IT assets. This API accepts a list of multiple configuration items.
}
\examples{

}
