% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.applicationdiscoveryservice_operations.R
\name{delete_applications}
\alias{delete_applications}
\title{Deletes a list of applications and their associations with configuration items}
\usage{
delete_applications(configurationIds)
}
\arguments{
\item{configurationIds}{Configuration ID of an application to be deleted.}
}
\description{
Deletes a list of applications and their associations with configuration items.
}
\examples{

}
