% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.applicationdiscoveryservice_operations.R
\name{delete_tags}
\alias{delete_tags}
\title{Deletes the association between configuration items and one or more tags}
\usage{
delete_tags(configurationIds, tags = NULL)
}
\arguments{
\item{configurationIds}{[required] A list of configuration items with tags that you want to delete.}

\item{tags}{Tags that you want to delete from one or more configuration items. Specify the tags that you want to delete in a \emph{key}-\emph{value} format. For example:

\code{{"key": "serverType", "value": "webServer"}}}
}
\description{
Deletes the association between configuration items and one or more tags. This API accepts a list of multiple configuration items.
}
\section{Accepted Parameters}{
\preformatted{delete_tags(
  configurationIds = list(
    "string"
  ),
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}
}

