% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.applicationdiscoveryservice_operations.R
\name{describe_continuous_exports}
\alias{describe_continuous_exports}
\title{Lists exports as specified by ID}
\usage{
describe_continuous_exports(
  exportIds = list(
    "string"
  ),
  maxResults = 123,
  nextToken = "string"
)
}
\arguments{
\item{exportIds}{The unique IDs assigned to the exports.}

\item{maxResults}{A number between 1 and 100 specifying the maximum number of continuous export descriptions returned.}

\item{nextToken}{The token from the previous call to \code{DescribeExportTasks}.}
}
\description{
Lists exports as specified by ID. All continuous exports associated with your user account can be listed if you call \code{DescribeContinuousExports} as is without passing any parameters.
}
\examples{

}
