% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.applicationdiscoveryservice_operations.R
\name{describe_export_configurations}
\alias{describe_export_configurations}
\title{\code{DescribeExportConfigurations} is deprecated}
\usage{
describe_export_configurations(
  exportIds = list(
    "string"
  ),
  maxResults = 123,
  nextToken = "string"
)
}
\arguments{
\item{exportIds}{A list of continuous export ids to search for.}

\item{maxResults}{A number between 1 and 100 specifying the maximum number of continuous export descriptions returned.}

\item{nextToken}{The token from the previous call to describe-export-tasks.}
}
\description{
\code{DescribeExportConfigurations} is deprecated.
}
\details{
Use instead \href{http://docs.aws.amazon.com/application-discovery/latest/APIReference/API_DescribeExportTasks.html}{DescribeExportTasks}.
}
\examples{

}
