% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.applicationdiscoveryservice_operations.R
\name{describe_import_tasks}
\alias{describe_import_tasks}
\title{Returns an array of import tasks for your account, including status information, times, IDs, the Amazon S3 Object URL for the import file, and more}
\usage{
describe_import_tasks(filters = NULL, maxResults = NULL,
  nextToken = NULL)
}
\arguments{
\item{filters}{An array of name-value pairs that you provide to filter the results for the \code{DescribeImportTask} request to a specific subset of results. Currently, wildcard values aren't supported for filters.}

\item{maxResults}{The maximum number of results that you want this request to return, up to 100.}

\item{nextToken}{The token to request a specific page of results.}
}
\description{
Returns an array of import tasks for your account, including status information, times, IDs, the Amazon S3 Object URL for the import file, and more.
}
\section{Accepted Parameters}{
\preformatted{describe_import_tasks(
  filters = list(
    list(
      name = "IMPORT_TASK_ID"|"STATUS"|"NAME",
      values = list(
        "string"
      )
    )
  ),
  maxResults = 123,
  nextToken = "string"
)
}
}

