% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.applicationdiscoveryservice_operations.R
\name{get_discovery_summary}
\alias{get_discovery_summary}
\title{Retrieves a short summary of discovered assets}
\usage{
get_discovery_summary()
}
\description{
Retrieves a short summary of discovered assets.
}
\details{
This API operation takes no request parameters and is called as is at the command prompt as shown in the example.
}
\examples{

}
