% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.applicationdiscoveryservice_package.R
\docType{package}
\name{paws.applicationdiscoveryservice-package}
\alias{paws.applicationdiscoveryservice}
\alias{paws.applicationdiscoveryservice-package}
\title{paws.applicationdiscoveryservice: AWS Application Discovery Service}
\description{
AWS Application Discovery Service helps you plan application
migration projects by automatically identifying servers, virtual
machines (VMs), software, and software dependencies running in your
on-premises data centers. Application Discovery Service also collects
application performance data, which can help you assess the outcome of
your migration. The data collected by Application Discovery Service is
securely retained in an AWS-hosted and managed database in the cloud.
You can export the data as a CSV or XML file into your preferred
visualization tool or cloud-migration solution to plan your migration.
For more information, see [AWS Application Discovery Service
FAQ](http://aws.amazon.com/application-discovery/faqs/).
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
