% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.applicationdiscoveryservice_operations.R
\name{start_continuous_export}
\alias{start_continuous_export}
\title{Start the continuous flow of agent's discovered data into Amazon Athena}
\usage{
start_continuous_export()
}
\description{
Start the continuous flow of agent's discovered data into Amazon Athena.
}
\examples{

}
