% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.applicationdiscoveryservice_operations.R
\name{start_data_collection_by_agent_ids}
\alias{start_data_collection_by_agent_ids}
\title{Instructs the specified agents or connectors to start collecting data}
\usage{
start_data_collection_by_agent_ids(
  agentIds = list(
    "string"
  )
)
}
\arguments{
\item{agentIds}{[required] The IDs of the agents or connectors from which to start collecting data. If you send a request to an agent/connector ID that you do not have permission to contact, according to your AWS account, the service does not throw an exception. Instead, it returns the error in the \emph{Description} field. If you send a request to multiple agents/connectors and you do not have permission to contact some of those agents/connectors, the system does not throw an exception. Instead, the system shows \code{Failed} in the \emph{Description} field.}
}
\description{
Instructs the specified agents or connectors to start collecting data.
}
\examples{

}
