% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.applicationdiscoveryservice_operations.R
\name{start_export_task}
\alias{start_export_task}
\title{Begins the export of discovered data to an S3 bucket}
\usage{
start_export_task(exportDataFormat = NULL, filters = NULL,
  startTime = NULL, endTime = NULL)
}
\arguments{
\item{exportDataFormat}{The file format for the returned export data. Default value is \code{CSV}. \strong{Note:} \emph{The} \code{GRAPHML} \emph{option has been deprecated.}}

\item{filters}{If a filter is present, it selects the single \code{agentId} of the Application Discovery Agent for which data is exported. The \code{agentId} can be found in the results of the \code{DescribeAgents} API or CLI. If no filter is present, \code{startTime} and \code{endTime} are ignored and exported data includes both Agentless Discovery Connector data and summary data from Application Discovery agents.}

\item{startTime}{The start timestamp for exported data from the single Application Discovery Agent selected in the filters. If no value is specified, data is exported starting from the first data collected by the agent.}

\item{endTime}{The end timestamp for exported data from the single Application Discovery Agent selected in the filters. If no value is specified, exported data includes the most recent data collected by the agent.}
}
\description{
Begins the export of discovered data to an S3 bucket.
}
\details{
If you specify \code{agentIds} in a filter, the task exports up to 72 hours of detailed data collected by the identified Application Discovery Agent, including network, process, and performance details. A time range for exported agent data may be set by using \code{startTime} and \code{endTime}. Export of detailed agent data is limited to five concurrently running exports.

If you do not include an \code{agentIds} filter, summary data is exported that includes both AWS Agentless Discovery Connector data and summary data from AWS Discovery Agents. Export of summary data is limited to two exports per day.
}
\section{Accepted Parameters}{
\preformatted{start_export_task(
  exportDataFormat = list(
    "CSV"|"GRAPHML"
  ),
  filters = list(
    list(
      name = "string",
      values = list(
        "string"
      ),
      condition = "string"
    )
  ),
  startTime = as.POSIXct("2015-01-01"),
  endTime = as.POSIXct("2015-01-01")
)
}
}

