% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.applicationdiscoveryservice_operations.R
\name{stop_continuous_export}
\alias{stop_continuous_export}
\title{Stop the continuous flow of agent's discovered data into Amazon Athena}
\usage{
stop_continuous_export(exportId)
}
\arguments{
\item{exportId}{[required] The unique ID assigned to this export.}
}
\description{
Stop the continuous flow of agent's discovered data into Amazon Athena.
}
\section{Accepted Parameters}{
\preformatted{stop_continuous_export(
  exportId = "string"
)
}
}

