% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.applicationdiscoveryservice_operations.R
\name{stop_data_collection_by_agent_ids}
\alias{stop_data_collection_by_agent_ids}
\title{Instructs the specified agents or connectors to stop collecting data}
\usage{
stop_data_collection_by_agent_ids(agentIds)
}
\arguments{
\item{agentIds}{[required] The IDs of the agents or connectors from which to stop collecting data.}
}
\description{
Instructs the specified agents or connectors to stop collecting data.
}
\section{Accepted Parameters}{
\preformatted{stop_data_collection_by_agent_ids(
  agentIds = list(
    "string"
  )
)
}
}

