% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.applicationdiscoveryservice_operations.R
\name{update_application}
\alias{update_application}
\title{Updates metadata about an application}
\usage{
update_application(
  configurationId = "string",
  name = "string",
  description = "string"
)
}
\arguments{
\item{configurationId}{[required] Configuration ID of the application to be updated.}

\item{name}{New name of the application to be updated.}

\item{description}{New description of the application to be updated.}
}
\description{
Updates metadata about an application.
}
\examples{

}
