% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.applicationdiscoveryservice_operations.R
\name{disassociate_configuration_items_from_application}
\alias{disassociate_configuration_items_from_application}
\title{Disassociates one or more configuration items from an application}
\usage{
disassociate_configuration_items_from_application(
  applicationConfigurationId = "string",
  configurationIds = list(
    "string"
  )
)
}
\arguments{
\item{applicationConfigurationId}{[required] Configuration ID of an application from which each item is disassociated.}

\item{configurationIds}{[required] Configuration ID of each item to be disassociated from an application.}
}
\description{
Disassociates one or more configuration items from an application.
}
\examples{

}
